+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Improved
+
+- Static files served by bedrock are now cached for an hour. Project resources are not cached, only references to `node_modules` and bedrock resources.
+
 ## 14.1.4 - 2024-03-27
 
 ### Fixed